==> OptionMetrics/2_Regressions_Generate.ipynb <==
import sys
import os
# Append the library path to PYTHONPATH, so library can be imported.
sys.path.append(os.path.dirname(os.getcwd()))

import shutil
import datetime
import numpy as np
import pandas as pd

from library import common as cm
from library import regression_aux as raux

%run setup.py
%matplotlib inline
%load_ext autoreload
%autoreload 2

np.random.seed(RANDOM_SEED)

%run Load_Clean_aux.py
max_period = max([int(s[6:]) for s in df.columns if 'period' in s])

print(f'The number of samples after loading is : {df.shape[0]}')

regr_dir = res_dir + 'Regression/'
print(f'Results are saved at:\n{regr_dir}')

if os.path.exists(regr_dir):
    shutil.rmtree(regr_dir)
os.makedirs(regr_dir)
shutil.copy('setup.py', regr_dir);

def save_tables(res, dirs):
    if res is not None:
        res['df_coef'].to_csv(f'{dirs}coef.csv')
        res['df_fit_std'].to_csv(f'{dirs}std.csv')

#### No hedge


sub_regr_dir = f'{regr_dir}No_Hedge/'
os.makedirs(f'{sub_regr_dir}pnl/', exist_ok=True)
pnl_path = f'{sub_regr_dir}pnl/pnl.csv'

zero = np.array([0.]*len(df))
zero = pd.Series(zero, index=df.index)

cm.store_pnl(df, zero, pnl_path)
save_tables(None, sub_regr_dir)

pnl = pd.read_csv(regr_dir + 'No_Hedge/pnl/pnl.csv', index_col=0)

#### Baseline: Black-Scholes 


sub_regr_dir = f'{regr_dir}BS_Benchmark/'
os.makedirs(f'{sub_regr_dir}pnl/', exist_ok=True)
pnl_path = f'{sub_regr_dir}pnl/pnl.csv'

cm.store_pnl(df, df['delta_bs'], pnl_path)
save_tables(None, sub_regr_dir)

#### Using fixed constants 0.9, 1.1


sub_regr_dir = f'{regr_dir}Fixed_Constants/'
os.makedirs(f'{sub_regr_dir}pnl/', exist_ok=True)
pnl_path = f'{sub_regr_dir}pnl/pnl.csv'

bl_c = df['cp_int'] == 0
delta = 0.9 * df['delta_bs']
delta[~bl_c] = 1.1 * df.loc[~bl_c, 'delta_bs']

cm.store_pnl(df, delta, pnl_path)
save_tables(None, sub_regr_dir)

#### Delta-only


sub_regr_dir = f'{regr_dir}Delta_only/'
kwargs = { 
      'features': ['delta_bs'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': False,
      'agg_side': False,
      'leverage': True}

res = raux.run_store_lin(**kwargs)
save_tables(res, dirs=sub_regr_dir)
res['df_leve'].to_csv(f'{sub_regr_dir}leverage.csv')

#### Delta-Vega


sub_regr_dir = f'{regr_dir}Delta_Vega/'
kwargs = {
      'features': ['delta_bs', 'vega_n'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': False,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Delta-Gamma


sub_regr_dir = f'{regr_dir}Delta_Gamma/'
kwargs = {
      'features': ['delta_bs', 'gamma_n'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': False,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Delta-Vanna


sub_regr_dir = f'{regr_dir}Delta_Vanna/'
kwargs = {
    'features': ['delta_bs', 'vanna_n'],
    'max_period': max_period,
    'sub_res': f'{sub_regr_dir}pnl/',
    'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
    'df': df,
    'delta_coeff_1': False
}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Delta-Gamma-Vanna


sub_regr_dir = f'{regr_dir}Delta_Gamma_Vanna/'
kwargs = {
    'features': ['delta_bs', 'gamma_n', 'vanna_n'],
    'max_period': max_period,
    'sub_res': f'{sub_regr_dir}pnl/',
    'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
    'df': df,
    'delta_coeff_1': False
}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Delta-Vega-Gamma


sub_regr_dir = f'{regr_dir}Delta_Vega_Gamma/'
kwargs = {
    'features': ['delta_bs', 'vega_n', 'gamma_n'],
    'max_period': max_period,
    'sub_res': f'{sub_regr_dir}pnl/',
    'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
    'df': df,
    'delta_coeff_1': False
}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Delta-Vega-Vanna


sub_regr_dir = f'{regr_dir}Delta_Vega_Vanna/'
kwargs = {
    'features': ['delta_bs', 'vega_n', 'vanna_n'],
    'max_period': max_period,
    'sub_res': f'{sub_regr_dir}pnl/',
    'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
    'df': df,
    'delta_coeff_1': False
}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Delta-Vega-Gamma-Vanna


sub_regr_dir = f'{regr_dir}Delta_Vega_Gamma_Va/'
kwargs = {
    'features': ['delta_bs', 'vega_n', 'gamma_n', 'vanna_n'],
    'max_period': max_period,
    'sub_res': f'{sub_regr_dir}pnl/',
    'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
    'df': df,
    'delta_coeff_1': False
}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Vega-only


sub_regr_dir = f'{regr_dir}Vega_only/'
kwargs = {
      'features': ['vega_n'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': True,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Vanna-only


sub_regr_dir = f'{regr_dir}Vanna_only/'
kwargs = {
      'features': ['vanna_n'],   
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': True,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Gamma-only


sub_regr_dir = f'{regr_dir}Gamma_only/'
kwargs = {
      'features': ['gamma_n'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': True,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Bias only


sub_regr_dir = f'{regr_dir}Bias/'
df['bias'] = 1.
kwargs = { 
      'features': ['bias'],   
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': True,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)

#### Hull-White


sub_regr_dir = f'{regr_dir}Hull_White/'
kwargs = {
      'features': ['vega_s', 'delta_vega_s', 'delta2_vega_s'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': True,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)


sub_regr_dir = f'{regr_dir}Hull_White_relaxed/'
kwargs = {
      'features': ['delta_bs', 'vega_s', 'delta_vega_s', 'delta2_vega_s'], 
      'max_period': max_period, 
      'sub_res': f'{sub_regr_dir}pnl/',
      'pnl_path': f'{sub_regr_dir}pnl/pnl.csv',
      'df': df,
      'delta_coeff_1': False,
      'agg_side': False}

res = raux.run_store_lin(**kwargs)
save_tables(res, sub_regr_dir)
