==> OptionMetrics/5_Diagnostic.ipynb <==
import sys
import os
# Append the library path to PYTHONPATH, so library can be imported.
sys.path.append(os.path.dirname(os.getcwd()))
import itertools
import datetime

import numpy as np
import pandas as pd

from library import common as cm

%run setup.py
%load_ext autoreload
%autoreload 2

# Cols are different data sets, rows are different experiments.
cols = ['FREQ=1D_HALFMONEY=otm_MINM=0.8_MAXM=1.5_MINTAU=0_MISSING=replace_missing_WINDOW=180D_KickTmrZeroVolume=False',
        'FREQ=2D_HALFMONEY=otm_MINM=0.8_MAXM=1.5_MINTAU=0_MISSING=replace_missing_WINDOW=180D_KickTmrZeroVolume=False',
       ]
rows = ['Regression/No_Hedge', 
        'Regression/BS_Benchmark', 
        'Regression/Fixed_Constants',
        'Regression/Delta_only', 'Regression/Vega_only', 'Regression/Gamma_only','Regression/Vanna_only',  
        'Regression/Bias',
        'Regression/Delta_Gamma', 'Regression/Delta_Vega',   'Regression/Delta_Vanna',
        'Regression/Delta_Vega_Gamma', 'Regression/Delta_Vega_Vanna', 'Regression/Delta_Gamma_Vanna',   
        'Regression/Delta_Vega_Gamma_Va', 
        'Regression/Hull_White', 'Regression/Hull_White_relaxed',
        #'Network/Normal_Feature', 
        'Network/Delta_Vega', 
        #'Network/Delta_Vega_Vanna',
        #'Network/Normal_Feature_CDF', 'Network/Delta_Vega_CDF',
        #'Network/Delta_Vega_Vanna_CDF',
        #'Network/Spot_Strike', 'Network/Spot_Strike_2'
]

def run_save_tables(dirs, agg, measure, op_type):
    inspector = cm.Inspector()
    df_res = inspector.evalPnls(df_dirs, aggregating=aggregating, measure=measure, op_type=op_type)
    if op_type == None:
        file = f'{DATA_DIR}Result/call_and_put_{measure}.csv'
    else:
        file = f'{DATA_DIR}Result/{op_type}_{measure}.csv'
    with open(file, 'w+') as f:
        f.write(f'{datetime.datetime.now()}\n{op_type}_{measure}\n')
    (df_res).to_csv(file, mode='a') 
    return df_res

df_dirs = pd.DataFrame(columns=cols, index=rows)
for x, y in list(itertools.product(rows, cols)):
    df_dirs.loc[x, y] = f'{DATA_DIR}Result/{y}/{x}/'

aggregating = 'mean'
measure = 'mse'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')

aggregating = 'mean'
measure = 'median'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='call')

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type='put')

aggregating = 'mean'
measure = 'lower5%VaR'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

aggregating = 'mean'
measure = 'upper95%VaR'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)

aggregating = 'mean'
measure = 'mean'

run_save_tables(df_dirs, agg=aggregating, measure=measure, op_type=None)
